ddsketch_agent

Struct DDSketch

pub struct DDSketch { /* private fields */ }

DDSketch implementation based on the Datadog Agent.

This implementation is subtly different from the open-source implementations of DDSketch, as Datadog made some slight tweaks to configuration values and in-memory layout to optimize it for insertion performance within the agent.

Weโve mimicked the agent version of DDSketch here in order to support a future where we can take sketches shipped by the agent, handle them internally, merge them, and so on, without any loss of accuracy, eventually forwarding them to Datadog ourselves.

As such, this implementation is constrained in the same ways: the configuration parameters cannot be changed, the collapsing strategy is fixed, and we support a limited number of methods for inserting into the sketch.

Importantly, we have a special function, again taken from the agent version, to allow us to interpolate histograms, specifically our own aggregated histograms, into a sketch so that we can emit useful default quantiles, rather than having to ship the buckets โ upper bound and count โ to a downstream system that might have no native way to do the same thing, basically providing no value as they have no way to render useful data from them.

Features

This crate exposes a single feature, serde, which enables serialization and deserialization of DDSketch with serde. This feature is not enabled by default, as it can be slightly risky to use. This is primarily due to the fact that the format of DDSketch is not promised to be stable over time. If you enable this feature, you should take care to avoid storing serialized DDSketch data for long periods of time, as deserializing it in the future may work but could lead to incorrect/unexpected behavior or issues with correctness.

Implementations

impl DDSketch

pub fn bin_count(&self) -> usize

Returns the number of bins in the sketch.

pub fn is_empty(&self) -> bool

Whether or not this sketch is empty.

pub fn count(&self) -> u32

Number of samples currently represented by this sketch.

pub fn min(&self) -> Option<f64>

Minimum value seen by this sketch.

Returns None if the sketch is empty.

pub fn max(&self) -> Option<f64>

Maximum value seen by this sketch.

Returns None if the sketch is empty.

pub fn sum(&self) -> Option<f64>

Sum of all values seen by this sketch.

Returns None if the sketch is empty.

pub fn avg(&self) -> Option<f64>

Average value seen by this sketch.

Returns None if the sketch is empty.

pub fn bins(&self) -> &[Bin]

Returns the current bins of this sketch.

pub fn clear(&mut self)

Clears the sketch, removing all bins and resetting all statistics.

pub fn insert(&mut self, v: f64)

Inserts a single value into the sketch.

pub fn insert_many(&mut self, vs: &[f64])

Inserts many values into the sketch.

pub fn insert_n(&mut self, v: f64, n: u32)

Inserts a single value into the sketch n times.

pub fn insert_interpolate_buckets( &mut self, buckets: Vec<Bucket>, ) -> Result<(), &'static str>

Errors

Returns an error if a bucket size is greater that u32::MAX.

pub fn quantile(&self, q: f64) -> Option<f64>

Gets the value at a given quantile.

pub fn merge(&mut self, other: &DDSketch)

Merges another sketch into this sketch, without a loss of accuracy.

All samples present in the other sketch will be correctly represented in this sketch, and summary statistics such as the sum, average, count, min, and max, will represent the sum of samples from both sketches.

pub fn merge_to_dogsketch(&self, dogsketch: &mut Dogsketch)

Merges this sketch into the Dogsketch Protocol Buffers representation.